requirements.txt
flask
pytest
pytest-cov
Pillow
